requirements.txt
google_cloud_aiplatform==1.70.0
opencv_python==4.10.0.84
Pillow==11.0.0
protobuf==5.28.3
python-dotenv==1.0.1
reportlab==4.2.5
vertexai==1.70.0
moviepy==1.0.3
